+++ apps/frontend/src/modules/auth/slices/auth.slice.ts
@@ -1,4 +1,4 @@
-import { createSlice, isPending, isRejected } from "@reduxjs/toolkit";
+import { createSlice } from "@reduxjs/toolkit";
 
 import { DataStatus } from "~/libs/enums/enums.js";
 import { type ValueOf } from "~/libs/types/types.js";
@@ -9,45 +9,66 @@ import { authPatch, getAuthenticatedUser, signIn, signUp } from "./actions.js";
 type State = {
 	authenticatedUser: null | UserAuthResponseDto;
 	dataStatus: ValueOf<typeof DataStatus>;
+	isInitialized: boolean;
 };
 
 const initialState: State = {
 	authenticatedUser: null,
 	dataStatus: DataStatus.IDLE,
+	isInitialized: false,
 };
 
 const { actions, name, reducer } = createSlice({
-	extraReducers: (builder) => {
-		builder
-			.addCase(getAuthenticatedUser.fulfilled, (state, action) => {
-				state.authenticatedUser = action.payload ? action.payload.data : null;
-				state.dataStatus = DataStatus.FULFILLED;
-			})
-			.addCase(signUp.fulfilled, (state, action) => {
-				state.authenticatedUser = action.payload.data.user;
-				state.dataStatus = DataStatus.FULFILLED;
-			})
-			.addCase(signIn.fulfilled, (state, action) => {
-				state.authenticatedUser = action.payload.data.user;
-				state.dataStatus = DataStatus.FULFILLED;
-			})
-			.addCase(authPatch.fulfilled, (state, action) => {
-				state.authenticatedUser = action.payload.data;
-				state.dataStatus = DataStatus.FULFILLED;
-			})
-			.addMatcher(
-				isPending(getAuthenticatedUser, signUp, signIn, authPatch),
-				(state) => {
-					state.dataStatus = DataStatus.PENDING;
-				},
-			)
-			.addMatcher(
-				isRejected(getAuthenticatedUser, signUp, signIn, authPatch),
-				(state) => {
-					state.authenticatedUser = null;
-					state.dataStatus = DataStatus.REJECTED;
-				},
-			);
+	extraReducers(builder) {
+		builder.addCase(getAuthenticatedUser.fulfilled, (state, action) => {
+			state.authenticatedUser = action.payload ? action.payload.data : null;
+			state.dataStatus = DataStatus.FULFILLED;
+			state.isInitialized = true;
+		});
+		builder.addCase(getAuthenticatedUser.pending, (state) => {
+			state.dataStatus = DataStatus.PENDING;
+		});
+		builder.addCase(getAuthenticatedUser.rejected, (state) => {
+			state.authenticatedUser = null;
+			state.dataStatus = DataStatus.REJECTED;
+			state.isInitialized = true;
+		});
+
+		builder.addCase(signUp.fulfilled, (state, action) => {
+			state.authenticatedUser = action.payload.data.user;
+			state.dataStatus = DataStatus.FULFILLED;
+		});
+		builder.addCase(signUp.pending, (state) => {
+			state.dataStatus = DataStatus.PENDING;
+		});
+		builder.addCase(signUp.rejected, (state) => {
+			state.authenticatedUser = null;
+			state.dataStatus = DataStatus.REJECTED;
+		});
+
+		builder.addCase(signIn.pending, (state) => {
+			state.dataStatus = DataStatus.PENDING;
+		});
+		builder.addCase(signIn.fulfilled, (state, action) => {
+			state.authenticatedUser = action.payload.data.user;
+			state.dataStatus = DataStatus.FULFILLED;
+		});
+		builder.addCase(signIn.rejected, (state) => {
+			state.authenticatedUser = null;
+			state.dataStatus = DataStatus.REJECTED;
+		});
+
+		builder.addCase(authPatch.fulfilled, (state, action) => {
+			state.authenticatedUser = action.payload.data;
+			state.dataStatus = DataStatus.FULFILLED;
+		});
+		builder.addCase(authPatch.pending, (state) => {
+			state.dataStatus = DataStatus.PENDING;
+		});
+		builder.addCase(authPatch.rejected, (state) => {
+			state.authenticatedUser = null;
+			state.dataStatus = DataStatus.REJECTED;
+		});
 	},
 	initialState,
 	name: "auth",