+++ packages/shared/src/modules/reviews/reviews.ts
@@ -1,0 +1,6 @@
+export {
+	type ReviewCreatePayload,
+	type ReviewGetByIdResponseDto,
+	type ReviewRequestDto,
+} from "./libs/types/types.js";
+export { reviewCreate as reviewCreateValidationSchema } from "./libs/validation-schemas/validation-schemas.js";